phase-4/backend/requirements.txt
# Core dependencies (from pyproject.toml)
alembic>=1.17.2
asyncpg>=0.31.0
better-auth>=0.0.1b11
openai-chatkit>=0.1.0
cryptography>=46.0.3
fastapi>=0.124.0
fastapi-cors>=0.0.6
fastmcp
httpx>=0.28.1
mcp>=1.25.0
openai>=2.14.0
openai-agents>=0.6.4
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pydantic-settings>=2.12.0
pyjwt>=2.10.1
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-cov>=7.0.0
python-dotenv>=1.2.1
python-jose>=3.5.0
pytz>=2025.2
sqlalchemy[asyncio]>=2.0.45
sqlmodel>=0.0.27
starlette>=0.50.0
uvicorn[standard]>=0.38.0
